backend/requirements.txt
fastapi==0.112.2
fastapi-cli==0.0.5
psycopg==3.1.19
psycopg-binary==3.1.19
pydantic==2.8.2
pydantic-settings==2.4.0
pytest==8.3.2
python-magic==0.4.27
requests==2.31.0
python-multipart==0.0.20
httpx==0.28.1